Extract status logic.
