+++ app/Models/Sprint.php
@@ -2,6 +2,7 @@
 
 use Phragile\PhabricatorAPI;
 use Phragile\TaskList;
+use Phragile\StatusByStatusFieldDispatcher;
 
 class Sprint extends Eloquent {
 
@@ -123,7 +124,7 @@ class Sprint extends Eloquent {
 
 		return [
 			'tasks' => $tasks,
-			'transactions' => $phabricator->getTaskTransactions((new TaskList($tasks))->getClosedTaskIDs())
+			'transactions' => $phabricator->getTaskTransactions((new TaskList($tasks, new StatusByStatusFieldDispatcher()))->getClosedTaskIDs())
 		];
 	}
