+++ app/Phragile/StatusDispatcher.php
@@ -1,0 +1,10 @@
+<?php
+namespace Phragile;
+
+interface StatusDispatcher {
+	/**
+	 * @param array $task
+	 * @return string $status
+	 */
+	public function getStatus(array $task);
+}
